refactor(user): migrate user ID type from int to int64

+++ internal/domain/user.go
@@ -3,7 +3,7 @@ package domain
 import "context"
 
 type User struct {
-	ID            int    `json:"id"`
+	ID            int64  `json:"id"`
 	Login         string `json:"login"`
 	Email         string `json:"email"`
 	Role          string `json:"role"`
@@ -17,8 +17,8 @@ type User struct {
 type UserRepository interface {
 	Create(ctx context.Context, user *User) error
 	GetByEmail(ctx context.Context, email string) (*User, error)
-	GetByID(ctx context.Context, id int) (*User, error)
+	GetByID(ctx context.Context, id int64) (*User, error)
 	GetAll(ctx context.Context) ([]*User, error)
 	Update(ctx context.Context, user *User) error
-	Delete(ctx context.Context, id int) error
+	Delete(ctx context.Context, id int64) error
 }
